+++ qiskit/providers/aer/pulse/de/qiskit_zvode.py
@@ -1,0 +1,34 @@
+# -*- coding: utf-8 -*-
+
+# This code is part of Qiskit.
+#
+# (C) Copyright IBM 2020.
+#
+# This code is licensed under the Apache License, Version 2.0. You may
+# obtain a copy of this license in the LICENSE.txt file in the root directory
+# of this source tree or at http://www.apache.org/licenses/LICENSE-2.0.
+#
+# Any modifications or derivative works of this code must retain this
+# copyright notice, and modified files need to carry a notice indicating
+# that they have been altered from the originals.
+
+# This file is part of QuTiP: Quantum Toolbox in Python.
+#
+#    Copyright (c) 2011 and later, Paul D. Nation and Robert J. Johansson.
+#    All rights reserved.
+# pylint: disable=no-value-for-parameter, invalid-name, import-error
+
+from scipy.integrate._ode import zvode
+
+class qiskit_zvode(zvode):
+    """Customized ZVODE with modified stepper so that
+    it always stops at a given time in tlist;
+    by default, it over shoots the time.
+    """
+    def step(self, *args):
+        itask = self.call_args[2]
+        self.rwork[0] = args[4]
+        self.call_args[2] = 5
+        r = self.run(*args)
+        self.call_args[2] = itask
+        return r
